2.1 Личный проект: валидируй, как никогда не валидировал
