--- src/components/app.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-import Main from "./main.jsx";
-
-const App = (props) => {
-  // eslint-disable-next-line react/prop-types
-  const {placesCount} = props;
-
-  return (
-    <Main placesCount={placesCount}/>
-  );
-};
-
-export default App;
